Allow explicitly setting `--color` and `--hyperlink` to "true if printing to tty"
Also known as `auto`; `auto` and `inverse` are meant to represent true/false but set as a default/without
an argument.
So:
  `f` == `f -c=auto`
	`f` == `f --hy=inverse`
and
  `f -c` == `f -c=inverse`
  `f --hy` == `f --hy=auto`

In `--hyperlink`'s case, `inverse` is the same as `false`; in fact in most cases this will hold true
and is probably why `auto` is seen in other programs while `inverse` is not.